requirements.txt
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
cloudinary==1.42.2
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==4.2
django-allauth==65.1.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-dotenv==1.4.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
pillow==11.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pyOpenSSL==25.0.0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.9.0
zope.interface==7.2
